+++ site-cookbooks/alterf/recipes/deploy.rb
@@ -6,6 +6,12 @@
 #
 # All rights reserved - Do Not Redistribute
 #
+node[:alterf][:requirements].each do |package_name|
+  package package_name do
+    action :install
+  end
+end
+
 deploy node[:alterf][:deploy_dir] do
   repo node[:alterf][:repository]
   branch ENV['ALTERF_VERSION'] || node[:alterf][:branch]
@@ -25,10 +31,6 @@ deploy node[:alterf][:deploy_dir] do
       end
     end
 
-    package 'mysql-devel' do
-      action :install
-    end
-
     link File.join(release_path, 'vendor/bundle') do
       to File.join(release_path, '../../shared/bundle')
       link_type :symbolic
@@ -37,6 +39,7 @@ deploy node[:alterf][:deploy_dir] do
     execute 'rvm 2.2.0 do bundle install --path=vendor/bundle' do
       cwd release_path
       environment 'PATH' => node[:rvm][:path]
+      not_if { node.chef_environment == 'compute' }
     end
   end
 end